requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
diploma-frontend @ file:///mnt/c/Users/Demodelife/PycharmProjects/Python_django_diploma_dpo/marketplace/diploma-frontend/dist/diploma-frontend-0.2.tar.gz
Django==4.1.7
django-credit-cards==0.4.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
docutils==0.19
drf-yasg==1.21.5
future==0.18.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-codec==1.3.2
packaging==23.0
Pillow==9.4.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
serializers==0.2.4
simplejson==3.19.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
